Add files via upload
Uno Game in Java. Executable through Linux terminal.
How to play:
The program will automatically shuffle the deck, ask for the numbers of players, and deal out decks of cards correspond to the number provided by the player. The player plays by typing the card number and card color following the rule of UNO. The first deck to empty out the hand wins.